Improve SourceCodeLocation accuracy
CssDeclarationNode did not have the SourceCodeLocation as a constructor
parameter, nor did it preserve it on a deep copy.

CssNodesListNode now tries to build its locations from its constituent nodes
unless a SourceCodeLocation is set explicitly.

Several bugs in the actual parser are also fixed as are numerous problems in
FixupFontDeclarations.

There are new tests and a SourceCodeLocationSubject for Truth which should
make new tests easier to write.

If this CL breaks your test, please fix your test instead of rolling back this CL:

• For [] conformance failures, add the offending file to the whitelist at
social/[]/ui/css_conformance.textproto. See CL 141453954 as an example.

• For failures where the test is checking node boundaries, you probably have
to just add 1 to the end boundary. This is because the compiler was
inconsistent in how it measured nodes.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=141544342