Introduce webfiles zip helpers
The web compiler will be using zip files in two capacities:

1. To store object files produced by each incremental compilation rule.
   Think of these .i.zip files (incremental zips) as like jars. They
   contain both the raw sources, and also possibly intermediate
   serialized AST cache variations. It might also contain magic files
   that the compiler pulled out from embedded <script> tags into their
   own files. These zips use light compression to favor speed. It is
   also smart enough to know which files are already compressed, e.g.
   jpeg, and not bother compressing them further.

2. To create deploy archives. Unlike incremental zips, deploy zips
   contain a minimal calculation of the transitive closure of assets
   which takes into consideration the links the compiler determined
   between files, along with things like runtime_deps edges. These
   zip files are optimized for heavy compression to favor small size.

Special care is taken of course to ensure the timestamps inherent in
the zip standard remain deterministic.

It is also worthing noting that the design introduced by this CL is
already deprecated. Bazel has a TreeArtifact feature which will replace
this system, once it's ready. After speaking with the Bazel team, they
communicated this will be a difficult API to design, and as a result,
are taking their time to make sure it's done right on the first try.

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=153208438